Convert settings to a package
This makes it possible to change and extend existing settings in
local.py without having to redefine them.

Also add a gitignore for local_code.py, which can be used to include
some local code such as an authentication backend. This can't be done in
local.py, as some classes cannot be imported until settings are
processed.