Expand Alexia API to include some methods for GDPR-compliance
These methods will mostly be used by Inter-Actief's website to facilitate that users are able to request a data-export of all their user data. Because this also includes personal data stored in other systems, and we want to provide a single place to create this export, we need a way to export this information from Alexia.